+++ cmake/Doxyfile.in
@@ -801,7 +801,7 @@ INPUT      = "@CMAKE_SOURCE_DIR@/include" \
 			 "@CMAKE_SOURCE_DIR@/include/math" \
 			 "@CMAKE_SOURCE_DIR@/include/Operator" \
 			 "@CMAKE_SOURCE_DIR@/include/OrbitalOptimization" \
-			 "@CMAKE_SOURCE_DIR@/include/properties" \
+			 "@CMAKE_SOURCE_DIR@/include/Properties" \
 			 "@CMAKE_SOURCE_DIR@/include/RDM" \
 			 "@CMAKE_SOURCE_DIR@/include/RHF" \
 			 "@CMAKE_SOURCE_DIR@/include/utilities" \
@@ -816,7 +816,7 @@ INPUT      = "@CMAKE_SOURCE_DIR@/include" \
 			 "@CMAKE_SOURCE_DIR@/src/HamiltonianParameters" \
 			 "@CMAKE_SOURCE_DIR@/src/math" \
 			 "@CMAKE_SOURCE_DIR@/src/OrbitalOptimization" \
-			 "@CMAKE_SOURCE_DIR@/src/properties" \
+			 "@CMAKE_SOURCE_DIR@/src/Properties" \
 			 "@CMAKE_SOURCE_DIR@/src/RDM" \
 			 "@CMAKE_SOURCE_DIR@/src/RHF" \
 			 "@CMAKE_SOURCE_DIR@/src/utilities" \
