+++ include/Properties/CMakeLists.txt
@@ -1,0 +1,5 @@
+target_sources(gqcp
+        PRIVATE
+            expectation_values.hpp
+            properties.hpp
+        )
