+++ include/gqcp.hpp
@@ -107,8 +107,8 @@
 #include "OrbitalOptimization/OrbitalRotationGenerators.hpp"
 #include "OrbitalOptimization/QCMethodNewtonOrbitalOptimizer.hpp"
 
-#include "properties/expectation_values.hpp"
-#include "properties/properties.hpp"
+#include "Properties/expectation_values.hpp"
+#include "Properties/properties.hpp"
 
 #include "RDM/BaseRDMBuilder.hpp"
 #include "RDM/BaseSpinUnresolvedRDMBuilder.hpp"
