+++ src/CMakeLists.txt
@@ -15,7 +15,7 @@ add_subdirectory(HamiltonianBuilder)
 add_subdirectory(HamiltonianParameters)
 add_subdirectory(math)
 add_subdirectory(OrbitalOptimization)
-add_subdirectory(properties)
+add_subdirectory(Properties)
 add_subdirectory(RDM)
 add_subdirectory(RHF)
 add_subdirectory(utilities)
