+++ src/HamiltonianParameters/AtomicDecompositionParameters.cpp
