+++ tests/CMakeLists.txt
@@ -18,7 +18,7 @@ add_subdirectory(HamiltonianParameters)
 add_subdirectory(math)
 add_subdirectory(Operator)
 add_subdirectory(OrbitalOptimization)
-add_subdirectory(properties)
+add_subdirectory(Properties)
 add_subdirectory(RDM)
 add_subdirectory(RHF)
 add_subdirectory(utilities)
