+++ tests/RDM/FCIRDMBuilder_test.cpp
@@ -25,7 +25,7 @@
 #include "CISolver/CISolver.hpp"
 #include "HamiltonianBuilder/FCI.hpp"
 #include "HamiltonianParameters/HamiltonianParameters.hpp"
-#include "properties/expectation_values.hpp"
+#include "Properties/expectation_values.hpp"
