+++ tests/RHF/constrained_RHF_test.cpp
@@ -22,7 +22,8 @@
 
 #include "RHF/DIISRHFSCFSolver.hpp"
 #include "HamiltonianParameters/HamiltonianParameters.hpp"
-#include "properties/expectation_values.hpp"
+#include "Properties/expectation_values.hpp"
+
 #include <random>
